requirements.txt



Flask==1.1.2
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==20.0.4
google-cloud-vision==2.0.0
google-cloud-storage==1.31.2
google-cloud-texttospeech
Wand==0.6.3
